Upgrade to version 0.2 of coramin.
This is done in preparation for the
implementation of _copy_relaxation_with_local_data,
which should allow SOCEdgeCuts to be copied in the
latest versions of coramin.

However, set_input now no longer takes a
persistent_solver parameter.
This is the norm for relaxations in coramin 0.2,
but is a breaking API change.